+++ .gitignore
@@ -29,3 +29,6 @@ Desktop.ini
 \#*\#
 Session.vim
 .netrwhist
+
+# Ignore node stuff
+node_modules
