+++ src/YeastLogic.ts
@@ -1,0 +1,152 @@
+export interface YeastyGoodness {
+    fed: number;
+    happy: number;
+    waiting: number;
+    hungry: number;
+    starving: number;
+    dead: number;
+  }
+  
+const constants: {
+    maturationRate: YeastyGoodness,
+    healthMultiplier: YeastyGoodness,
+    hungerMultiplier: YeastyGoodness,
+    volumeMultiplier: YeastyGoodness,
+} = {
+    maturationRate: {
+        fed: 0.05,
+        happy: 0.05,
+        waiting: 0.01,
+        hungry: 0.01,
+        starving: 0.005,
+        dead: 0.0,
+      },
+    healthMultiplier: {
+        fed : 1.0,
+        happy : 1.0,
+        waiting : 1.0,
+        hungry : 0.75,
+        starving : 0.25,
+        dead : 0.0,
+      },
+    hungerMultiplier: {
+        fed : 0.0,
+        happy : 0.1,
+        waiting : 0.5,
+        hungry : 0.9,
+        starving : 1.0,
+        dead : 0.0,    
+    },
+    volumeMultiplier: {
+            fed : 1.0,
+            happy : 1.75,
+            waiting : 2.5,
+            hungry : 1.75,
+            starving : 1.25,
+            dead : 1.0,
+    }
+}
+  
+
+export function hunger(yeast: YeastyGoodness): number {
+    const alive = yeast.fed + yeast.happy + yeast.waiting + yeast.hungry + yeast.starving;
+    const hungerAbsolute = yeast.fed * constants.hungerMultiplier.fed  
+      + yeast.happy * constants.hungerMultiplier.happy
+      + yeast.waiting * constants.hungerMultiplier.waiting
+      + yeast.hungry * constants.hungerMultiplier.hungry
+      + yeast.starving * constants.hungerMultiplier.starving;
+    return hungerAbsolute / alive;
+  }
+
+  export function health(yeast: YeastyGoodness): number {
+    const total = yeast.fed + yeast.happy + yeast.waiting + yeast.hungry + yeast.starving + yeast.dead;
+    const hungerAbsolute = yeast.fed * constants.healthMultiplier.fed  
+      + yeast.happy * constants.healthMultiplier.happy
+      + yeast.waiting * constants.healthMultiplier.waiting
+      + yeast.hungry * constants.healthMultiplier.hungry
+      + yeast.starving * constants.healthMultiplier.starving;
+    return hungerAbsolute / total;
+  }
+  
+
+  export  function stepYeast(yeast: YeastyGoodness): YeastyGoodness {
+    const gotHappy = yeast.fed * constants.maturationRate.fed;
+    const gotWaiting = yeast.happy * constants.maturationRate.happy;
+    const gotHungry = yeast.waiting * constants.maturationRate.waiting;
+    const gotStarving = yeast.hungry * constants.maturationRate.hungry;
+    const gotDead = yeast.starving * constants.maturationRate.starving;
+  
+    return {
+      fed : yeast.fed - gotHappy,
+      happy : gotHappy + yeast.happy - gotWaiting,
+      waiting : gotWaiting + yeast.waiting - gotHungry,
+      hungry : gotHungry + yeast.hungry - gotStarving,
+      starving : gotStarving + yeast.starving - gotDead,
+      dead : gotDead + yeast.dead,
+    }
+  }
+  
+  export  function feedYeast(amount: number, yeast: YeastyGoodness): YeastyGoodness {
+    const feedMeSeymor = yeast.waiting + yeast.hungry + yeast.starving;
+    if (feedMeSeymor >= amount) {
+      // Not enough food to feed all the really hungry yeast
+      console.log("A");
+      return {
+        fed : yeast.fed + amount + amount,
+        happy : yeast.happy,
+        waiting : yeast.waiting - (yeast.waiting / feedMeSeymor) * amount,
+        hungry : yeast.hungry - (yeast.hungry / feedMeSeymor) * amount,
+        starving : yeast.starving - (yeast.starving / feedMeSeymor) * amount,
+        dead : yeast.dead,
+      };
+    } else if (feedMeSeymor + yeast.happy >= amount) {
+      // Feeding all the really hungry yeast, plus some the yeast that could have a bite
+      console.log("B");
+      return {
+        fed : yeast.fed + amount + amount,
+        happy : yeast.happy - (amount - feedMeSeymor),
+        waiting : 0,
+        hungry : 0,
+        starving : 0,
+        dead : yeast.dead,
+      };
+    } else {
+      // Some feed goes to waste as dead material
+      console.log("C");
+      return {
+        fed : (yeast.fed + yeast.happy + yeast.waiting + yeast.hungry + yeast.starving) * 2,
+        happy : 0,
+        waiting : 0,
+        hungry : 0,
+        starving : 0,
+        dead : yeast.dead + amount - (yeast.happy + yeast.waiting + yeast.hungry + yeast.starving),
+      };
+    }
+  }
+  
+export  function yeastVolume(yeast: YeastyGoodness): number {
+    return yeast.fed * constants.volumeMultiplier.fed
+    + yeast.happy * constants.volumeMultiplier.happy
+    + yeast.waiting * constants.volumeMultiplier.waiting
+    + yeast.hungry * constants.volumeMultiplier.hungry
+    + yeast.starving * constants.volumeMultiplier.starving
+    + yeast.dead * constants.volumeMultiplier.dead;
+  } 
+  
+  export  function clampYeast(maxVolume: number, yeast: YeastyGoodness): [YeastyGoodness, number] {
+    const volume = yeastVolume(yeast);
+    if (volume <= maxVolume) return [yeast, 0];
+    const loss = (volume - maxVolume)/volume;
+    return [
+      {
+        fed: yeast.fed * (1-loss),
+        happy: yeast.happy * (1-loss),
+        waiting: yeast.waiting * (1-loss),
+        hungry: yeast.hungry * (1-loss),
+        starving: yeast.starving * (1-loss),
+        dead: yeast.dead * (1-loss),
+      },
+      volume - maxVolume,
+    ]
+  }
+
